+++ NEWS.md
@@ -1,5 +1,6 @@
 # spant 4.2.0
 * Added rm_basis_elements function.
+* Added plot_tqn_fit_csv function.
 
 # spant 4.1.0
 * Update to lsei.f to fix a warning produced by GCC 16.