+++ NEWS.md
@@ -1,3 +1,6 @@
+# spant 4.3.0
+* Bug fix for mr_data2bids function.
+
 # spant 4.2.0
 * Added rm_basis_elements function.
 * Added plot_tqn_fit_csv function.